Fix logic for creating installments for a reservation
Previously we checked that if the reservation is in a state "reserved".
This does not work in e.g. the case being that offer has been accepted
and the salesperson wants to create installments for the customer. Then
the reservation is not in the reserved state and creating installments
is not possible. Change this logic to allow creating installments when
the reservation is not in the review state and the reservation is
first in the queue.

Also show dimmed text for installment rows where the reservation is
canceled but installments have already been saved.