Fix tests after updating packages
Our mock data can be a bit different from the actual data, e.g.
only have a portion of the values in that type. We need to cast
the proper types for dummy/mock data used to fix eslint errors with
the new eslint version.

Also remove some of the smoke tests that only render the component,
as we also check the actual component content in other tests. These
tests don't add any value so they are quite redundant and can be
removed.